Refactored ui menu
Refactored the entire menu scene, seperated logic into a main menu controller, and wrapped all ui elements in a UI main menu controller. Every canvas is organized and is contained in their own scripts with events for each button press.

Added the functionality of pressing esc to exit a canvas, and they do it in the right order using a stack.