Obstacles on the turn are now animated moving along the arc of the conveyor belt
The obstacles will not a) come in from the right of the belt after the turn b) turn with the arc of the belt improving the illusion of the conveyor belt and c) reset their transform to their original position after they have passed their turn as to not screw up any other components.

The rotations do look weird tho, so i will have to come up with a solution to turn the obstacles correctly as they are making the turn. Anyways, I am very pleased with the outcome so far.