Animated obstacles turning on the conveyor belt (#8)
* first commit

* made hand move from hidden to idle faster

* Added turn animation

Obstacles are now animated with a sharp 90 degree turn where the tunnel turns, and although its not perfect, its a lot better than how it used to look. Further implementation on this is needed and it would be nice if they made a circular turn instead.

* removed obstacle children with obstacle component

The obstacle component was causing bugs with the new turning animations, because the component that moves the obstacles looks for Obstacle components within a set, and therefore when certain plates with toppings was moved it looked very weird

* Messed around to figure out mathmatics of circular turn by designing gizmos

* Obstacles on the turn are now animated moving along the arc of the conveyor belt

The obstacles will not a) come in from the right of the belt after the turn b) turn with the arc of the belt improving the illusion of the conveyor belt and c) reset their transform to their original position after they have passed their turn as to not screw up any other components.

The rotations do look weird tho, so i will have to come up with a solution to turn the obstacles correctly as they are making the turn. Anyways, I am very pleased with the outcome so far.

* Some last changes to sets to adapt to the new turning belt,

There might be more, but this looks good enough to push and merge

* Obstacles now rotate correctly when making the conveyor belt turn

* Refactored a little

* enabled player collider